Add Q24-CircleArea.sql to calculate the area of a circle; add Q25-CirclePerimeter.sql to calculate the perimeter of a circle; replace Q23TriangleArea.sql with updated version

+++ DBMS-2/Assignment_1-2/Q23-TriangleArea.sql
@@ -8,7 +8,7 @@ DECLARE
     b NUMBER := &b;
     c NUMBER := &c;
     s NUMBER := (a + b + c) / 2;
-    area FLOAT;
+    area NUMBER(7,2);
 BEGIN
     area := SQRT(s * (s - a) * (s - b) * (s - c));
     DBMS_OUTPUT.PUT_LINE('The area is: ' || area);
